Let people without Deno format the docs (#325)
This is done by adding the `deno-bin` npm package to the development dependencies. That way, everyone who runs `npm install` will also get a local Deno binary in their `node_modules` folder. This can in turn be used to format the docs without global Deno installation on the system.

Given the size of `node_modules`, this will hardly be noticed. Just TypeScript has almost twice the size of the Deno binary.