Add note about specifying file extension.
I think it would be better to add this note to inform people about the path of file.
First time I looked at this I passed the path of a directory and not a file path and got `Illegal operation` error. Some people asked a solution for this and they were giving a directory path too.
P.S. I can't say I'm very well at English so you can change the sentences I added but including these as a note in documentation would probably benefit others and prevent getting error just because of being careless.